feat: add default directive
