+++ .gitignore
@@ -14,3 +14,6 @@
 # Dependency directories (remove the comment below to include it)
 # vendor/
 .vscode
+
+# Project specified
+.DS_Store
