+++ form.go
@@ -1,3 +1,6 @@
+// directive: "form"
+// https://ggicci.github.io/httpin/directives/form
+
 package httpin
 
 // formValueExtractor implements the "form" executor who extracts values from
