+++ gochi.go
@@ -1,4 +1,5 @@
-// URLParam extension for github.com/go-chi/chi.
+// integration: "gochi"
+// https://ggicci.github.io/httpin/integrations/gochi
 
 package httpin
 
@@ -10,6 +11,15 @@ import (
 // GochiURLParamFunc is chi.URLParam
 type GochiURLParamFunc func(r *http.Request, key string) string
 
+// UseGochiURLParam registers a directive executor which can extract values
+// from `chi.URLParam`, i.e. path variables.
+// https://ggicci.github.io/httpin/integrations/gochi
+//
+// Usage:
+//
+//     func init() {
+//         httpin.UseGochiURLParam("path", chi.URLParam)
+//     }
 func UseGochiURLParam(executor string, fn GochiURLParamFunc) {
 	RegisterDirectiveExecutor(executor, &gochiURLParamExtractor{URLParam: fn}, nil)
 }
